+++ packages/exception/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @open-dpp/exception
 
+## 2.6.0
+
+## 2.5.0
+
 ## 2.4.0
 
 ### Minor Changes